Closes #4 Avatar management Extract to avatar api dao
