+++ src/main/java/com/faforever/moderatorclient/api/domain/AvatarService.java
@@ -1,8 +1,7 @@
 package com.faforever.moderatorclient.api.domain;
 
 import com.faforever.commons.api.dto.Avatar;
-import com.faforever.moderatorclient.api.ElideRouteBuilder;
-import com.faforever.moderatorclient.api.FafApiCommunicationService;
+import com.faforever.moderatorclient.api.AvatarApiClient;
 import lombok.extern.slf4j.Slf4j;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Service;
@@ -13,69 +12,48 @@ import java.util.List;
 @Service
 @Slf4j
 public class AvatarService {
-    private final FafApiCommunicationService fafApi;
+    private final AvatarApiClient avatarApiClient;
 
-    public AvatarService(FafApiCommunicationService fafApi) {
-        this.fafApi = fafApi;
+    public AvatarService(AvatarApiClient avatarApiClient) {
+        this.avatarApiClient = avatarApiClient;
     }
 
     public List<Avatar> getAll() {
         log.debug("Retrieving all avatars");
-        List<Avatar> result = fafApi.getAll(ElideRouteBuilder.of(Avatar.class)
-                .addInclude("assignments")
-                .addInclude("assignments.player"));
-        log.trace("found {} avatars", result.size());
-        return result;
-    }
-
-    private List<Avatar> findAvatarsByAttribute(@NotNull String attribute, @NotNull String pattern) {
-        log.debug("Searching for avatars by attribute '{}' with pattern: {}", attribute, pattern);
-        ElideRouteBuilder<Avatar> routeBuilder = ElideRouteBuilder.of(Avatar.class)
-                .addInclude("assignments")
-                .addInclude("assignments.player")
-                .filter(ElideRouteBuilder.qBuilder().string(attribute).eq(pattern));
-
-        List<Avatar> result = fafApi.getAll(routeBuilder);
+        List<Avatar> result = avatarApiClient.getAllAvatarsWithPlayerAssignments();
         log.trace("found {} avatars", result.size());
         return result;
     }
 
     public List<Avatar> findAvatarsById(@NotNull String pattern) {
-        return findAvatarsByAttribute("id", pattern);
+        return avatarApiClient.findAvatarsById(pattern);
     }
 
     public List<Avatar> findAvatarsByTooltip(@NotNull String pattern) {
-        return findAvatarsByAttribute("tooltip", pattern);
+        return avatarApiClient.findAvatarsByTooltip(pattern);
     }
 
     public List<Avatar> findAvatarsByAssignedUser(@NotNull String pattern) {
         log.debug("Searching for avatars by assigned player with pattern: {}", pattern);
-        ElideRouteBuilder<Avatar> routeBuilder = ElideRouteBuilder.of(Avatar.class)
-                .addInclude("assignments")
-                .addInclude("assignments.player")
-                .filter(ElideRouteBuilder.qBuilder().string("assignments.player.id").eq(pattern)
-                        .or().string("assignments.player.login").eq(pattern));
-
-        List<Avatar> result = fafApi.getAll(routeBuilder);
+        List<Avatar> result = avatarApiClient.findAvatarsByAssignedUser(pattern);
         log.trace("found {} avatars", result.size());
         return result;
     }
 
     public void createAvatar(String name, File avatarImageFile) {
-        fafApi.uploadAvatar(name, avatarImageFile);
+        avatarApiClient.uploadAvatar(name, avatarImageFile);
     }
 
     public void updateAvatar(String avatarId, String name, File avatarImageFile) {
         if (avatarImageFile != null) {
-            fafApi.reuploadAvatar(avatarId, name, avatarImageFile);
+            avatarApiClient.reuploadAvatar(avatarId, name, avatarImageFile);
         } else {
-            fafApi.patch(ElideRouteBuilder.of(Avatar.class).id(avatarId),
-                    (Avatar) new Avatar().setTooltip(name).setId(avatarId));
+            avatarApiClient.updateAvatarMetadata(avatarId, name);
         }
 
     }
 
     public void deleteAvatar(String avatarId) {
-        fafApi.deleteAvatar(avatarId);
+        avatarApiClient.deleteAvatar(avatarId);
     }
 }
